src/python/converter/requirements.txt
Pillow==11.1.0
dearpygui==2.1.1
